Add notify-send notifier
Add journald-notify is meant to be run as a daemon in a systemd service
as a separate user, there's a good chance you'll need to use
notify-send-headless (provided by the proc pypi package) with sudo
(which also means you'll need to give the journald-notify user
permission to run this command with sudo without a password). This is
the default setup when using this notifier, to avoid unnecessary
configuration.